LPS-70229 Moving ContentTransformerContentTypes to the API module
