+++ adaptive-media-asset-publisher-web/build.gradle
@@ -8,5 +8,4 @@ dependencies {
 	provided group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	provided project(":apps:adaptive-media:adaptive-media-content-transformer-api")
-	provided project(":apps:adaptive-media:adaptive-media-image-content-transformer")
 }
