+++ adaptive-media-asset-publisher-web/src/main/java/com/liferay/adaptive/media/asset/publisher/web/internal/filter/AssetPublisherPortletFilter.java
@@ -15,7 +15,7 @@
 package com.liferay.adaptive.media.asset.publisher.web.internal.filter;
 
 import com.liferay.adaptive.media.content.transformer.ContentTransformerHandler;
-import com.liferay.adaptive.media.image.content.transformer.constants.ContentTransformerContentTypes;
+import com.liferay.adaptive.media.content.transformer.constants.ContentTransformerContentTypes;
 import com.liferay.asset.publisher.web.constants.AssetPublisherPortletKeys;
 import com.liferay.portal.kernel.servlet.BufferCacheServletResponse;
 import com.liferay.portal.kernel.servlet.ServletResponseUtil;
