+++ adaptive-media-blogs-web/src/main/java/com/liferay/adaptive/media/blogs/web/internal/filter/BlogsPortletFilter.java
@@ -15,7 +15,7 @@
 package com.liferay.adaptive.media.blogs.web.internal.filter;
 
 import com.liferay.adaptive.media.content.transformer.ContentTransformerHandler;
-import com.liferay.adaptive.media.image.content.transformer.constants.ContentTransformerContentTypes;
+import com.liferay.adaptive.media.content.transformer.constants.ContentTransformerContentTypes;
 import com.liferay.blogs.web.constants.BlogsPortletKeys;
 import com.liferay.portal.kernel.servlet.BufferCacheServletResponse;
 import com.liferay.portal.kernel.servlet.ServletResponseUtil;
