+++ adaptive-media-content-transformer-api/bnd.bnd
@@ -1,6 +1,8 @@
 Bundle-Name: Liferay Adaptive Media Content Transformer API
 Bundle-SymbolicName: com.liferay.adaptive.media.content.transformer.api
 Bundle-Version: 1.0.0
-Export-Package: com.liferay.adaptive.media.content.transformer
+Export-Package:\
+	com.liferay.adaptive.media.content.transformer,\
+	com.liferay.adaptive.media.content.transformer.constants
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title: Adaptive Media
